Add inital draft of MCollective PE upgrader
This only supports upgrading PE to PE, since I couldn't think of a good
way to get the install dir for mco agents on POSS. This can be mitigated
by specifying the `mco_agent_dir` parameter.

This is currently extremely limited and hardly tested, so I don't
recommend merging it just yet, but I'd love to get some love on it.

Note: this will likely never upgrade PE2.x to PE3.x.