Fix distro version resolution for Solaris 10
Previously, the pe_upgrade_distro_version fact used the `operatingsystemrelease`
fact to calculate the major distro version. In the tests, it expects to find this
fact in the format of '5.10', but it's not: kernelrelease is the fact that reports
that way. The operatingsystemrelease fact is formatted as '10u10'.

the case:

```
[root@ebstst0app1 ~]# facter -p operatingsystemrelease
operatingsystemrelease => 10_u10
[root@ebstst0app1 ~]# facter -p kernelrelease
operatingsystemrelease => 5.10
```

Because of this, the regex doesn't match correctly and sets the
pe_upgrade_distro_version fact as '10_u10' instead of '10'. This commit
updates the fact to use kernelrelease.